refactor: select 관련 라우팅을 일괄 query string으로 일괄 변경
