+++ src/app/card/(components)/go-card-benefit.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 export default function GoCardBenefit() {
   return (
     <Link
-      href={PATH.CARD_BENEFIT_DEFAULT}
+      href={PATH.CARD_BENEFIT}
       className="relative rounded-full bg-[#AB92FA] flex justify-between items-center p-6"
     >
       <h1 className="card-title text-black">더 큰 혜택 찾기</h1>
