+++ src/lib/_shared/paths.ts
@@ -7,9 +7,7 @@ export const PATH = {
 
   // in home
   PAYMENTS: "/payments",
-  PAYMENTS_DEFAULT: "/payments/all?sort=recent",
   CARD_BENEFIT: "/card/benefit",
-  CARD_BENEFIT_DEFAULT: "/card/benefit/traffic",
 
   // in fundraisings
   FUNDRAISINGS_LIST: "/fundraisings/list",
